requirements.txt
streamlit
langchain-openai
langchain-core
openai
python-dotenv
pytest